Add functions folder and add the generic functions to there
