+++ src/Functions/GetDirFromAngle.js
@@ -1,0 +1,11 @@
+const GetDirFromAngle = (angle) => {
+  const rads = angle * (Math.PI / 180);
+  const tx = Math.cos(rads);
+  const ty = Math.sin(rads);
+  return {
+    tx,
+    ty,
+  };
+};
+
+export default GetDirFromAngle;
