+++ src/Functions/GetTimer.js
@@ -1,0 +1,6 @@
+const GetTimer = () => {
+  const date = new Date();
+  return date.getTime();
+};
+
+export default GetTimer;
