+++ src/Functions/TimeClearer.js
@@ -1,0 +1,5 @@
+const TimeClearer = () => {
+  window.model.clock = 0;
+};
+
+export default TimeClearer;
