+++ src/Functions/ToDegres.js
@@ -1,0 +1,3 @@
+const ToDegres = (angle) => angle * (180 / Math.PI);
+
+export default ToDegres;
