+++ src/Functions/Toggler.js
@@ -1,0 +1,11 @@
+const Toggler = () => {
+  if (window.model.musicMode) {
+    window.model.backgroundMusic.stop();
+    window.model.musicMode = false;
+  } else {
+    window.model.backgroundMusic.play();
+    window.model.musicMode = true;
+  }
+};
+
+export default Toggler;
