+++ src/Objects/Button.js
@@ -18,14 +18,6 @@ export default class Button extends Phaser.GameObjects.Container {
       this.scene.scene.start(targetScene);
     });
 
-    this.button.on('pointerover', () => {
-      this.button.setTexture(key2);
-    });
-
-    this.button.on('pointerout', () => {
-      this.button.setTexture(key1);
-    });
-
     this.scene.add.existing(this);
   }
 }
